requirements.txt
Flask
google-generativeai
PyPDF2
